Renamed common project
