+++ fiskaltrust.Middleware.Interface.sln
@@ -26,9 +26,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "fiskaltrust.Middleware.Inte
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "fiskaltrust.Middleware.Interface.Client.Soap", "src\fiskaltrust.Middleware.Interface.Client.Soap\fiskaltrust.Middleware.Interface.Client.Soap.csproj", "{FB69506F-4B31-4F9E-90EC-078D035A17EB}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "fiskaltrust.Middleware.Interface.Client.Shared", "src\fiskaltrust.Middleware.Interface.Client.Shared\fiskaltrust.Middleware.Interface.Client.Shared.csproj", "{4CF9D784-E218-414D-8A14-56D514C0FBA9}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "fiskaltrust.Middleware.Interface.Client.Tests", "test\fiskaltrust.Middleware.Interface.Client.Tests\fiskaltrust.Middleware.Interface.Client.Tests.csproj", "{8D5E806B-39EF-4FCD-AD59-5AB710F5D8BB}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "fiskaltrust.Middleware.Interface.Client.Tests", "test\fiskaltrust.Middleware.Interface.Client.Tests\fiskaltrust.Middleware.Interface.Client.Tests.csproj", "{8D5E806B-39EF-4FCD-AD59-5AB710F5D8BB}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "fiskaltrust.Middleware.Interface.Client.Common", "src\fiskaltrust.Middleware.Interface.Client.Common\fiskaltrust.Middleware.Interface.Client.Common.csproj", "{4250D63B-366A-45C0-A2FB-0F9818961602}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -56,14 +56,14 @@ Global
 		{FB69506F-4B31-4F9E-90EC-078D035A17EB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FB69506F-4B31-4F9E-90EC-078D035A17EB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FB69506F-4B31-4F9E-90EC-078D035A17EB}.Release|Any CPU.Build.0 = Release|Any CPU
-		{4CF9D784-E218-414D-8A14-56D514C0FBA9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4CF9D784-E218-414D-8A14-56D514C0FBA9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4CF9D784-E218-414D-8A14-56D514C0FBA9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4CF9D784-E218-414D-8A14-56D514C0FBA9}.Release|Any CPU.Build.0 = Release|Any CPU
 		{8D5E806B-39EF-4FCD-AD59-5AB710F5D8BB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{8D5E806B-39EF-4FCD-AD59-5AB710F5D8BB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8D5E806B-39EF-4FCD-AD59-5AB710F5D8BB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8D5E806B-39EF-4FCD-AD59-5AB710F5D8BB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4250D63B-366A-45C0-A2FB-0F9818961602}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4250D63B-366A-45C0-A2FB-0F9818961602}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4250D63B-366A-45C0-A2FB-0F9818961602}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4250D63B-366A-45C0-A2FB-0F9818961602}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -75,8 +75,8 @@ Global
 		{5274E790-A7AD-4360-8613-72E9CCDF43CE} = {7CCEF987-3E5A-48D0-9D69-708C62526829}
 		{71C15B88-ACE8-4926-AA1D-85ECA2199AE3} = {7CCEF987-3E5A-48D0-9D69-708C62526829}
 		{FB69506F-4B31-4F9E-90EC-078D035A17EB} = {7CCEF987-3E5A-48D0-9D69-708C62526829}
-		{4CF9D784-E218-414D-8A14-56D514C0FBA9} = {7CCEF987-3E5A-48D0-9D69-708C62526829}
 		{8D5E806B-39EF-4FCD-AD59-5AB710F5D8BB} = {4A4977D1-0B9A-44F9-B4A3-3D08FBB58D14}
+		{4250D63B-366A-45C0-A2FB-0F9818961602} = {7CCEF987-3E5A-48D0-9D69-708C62526829}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {3CAE7E81-C6A5-44FA-8E98-F9207561129F}
