+++ src/fiskaltrust.Middleware.Interface.Client.Common/RetryLogic/Interfaces/IRetryPolicyHandler.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Threading.Tasks;
 
-namespace fiskaltrust.Middleware.Interface.Client.Shared.RetryLogic.Interfaces
+namespace fiskaltrust.Middleware.Interface.Client.Common.RetryLogic
 {
     public interface IRetryPolicyHandler<T>
     {
