+++ src/fiskaltrust.Middleware.Interface.Client.Http/HttpPosFactory.cs
@@ -1,5 +1,5 @@
 ﻿using fiskaltrust.ifPOS.v1;
-using fiskaltrust.Middleware.Interface.Client.Shared;
+using fiskaltrust.Middleware.Interface.Client.Common.RetryLogic;
 using System.Threading.Tasks;
 
 namespace fiskaltrust.Middleware.Interface.Client.Http
